Still working on deferred rendering. Some name changes and other stuff

+++ shader/geometry.frag
@@ -7,7 +7,7 @@ in VARYINGS {
 
 struct MaterialParameter {
 	vec3 diffuse_color;
-	float shininess;
+	int shininess;
 };
 
 uniform MaterialParameter material;
